requirements.txt
streamlit
python-dotenv
PyMuPDF             # For extracting text and images from PDFs
pdf2image        # To convert PDF pages to images
easyocr     # For OCR on extracted images
Pillow               # For handling images (PIL library)
numpy
python-pptx==0.6.21
pymongo
streamlit_extras